Moved rules to rules.universal.
