+++ idd3/rules/__init__.py
@@ -14,77 +14,5 @@
 #
 # You should have received a copy of the GNU General Public License along with
 # this program.  If not, see <http://www.gnu.org/licenses/>.
-
-from __future__ import print_function, unicode_literals, division
-
-# Atomic rulesets.
-from idd3.rules.atomic_rulesets import *
-
-# Atomic emitting rulesets.
-from idd3.rules.atomic_emitting_rulesets import *
-
-# Noun-Phrase rulesets.
-from idd3.rules.np_rulesets import *
-
-# Verb-Phrase rulesets.
-from idd3.rules.vp_rulesets import *
-
-# Adjectival-Phrase rulesets
-from idd3.rules.adjp_rulesets import *
-
-# Adverbial-Phrase rulesets
-from idd3.rules.advp_rulesets import *
-
-# Uncategorized rulesets.
-from idd3.rules.misc_rulesets import *
-
-all_rulesets = [TopRuleset(),
-                # Atomic rulesets.
-                PrtRuleset(),
-                AuxRuleset(),
-                AuxpassRuleset(),
-                CcRuleset(),
-                CopRuleset(),
-                ComplmRuleset(),
-                AdpRuleset(),
-                NumberRuleset(),
-                PreconjRuleset(),
-                MarkRuleset(),
-                PredetRuleset(),
-                # Atomic emitting rulesets.
-                NegRuleset(),
-                DiscourseRuleset(),
-                # Noun-Phrase rulesets.
-                NsubjRuleset(),
-                NsubjpassRuleset(),
-                DobjRuleset(),
-                AdpobjRuleset(),
-                IobjRuleset(),
-                PossRuleset(),
-                NmodRuleset(),
-                TmodRuleset(),
-                ApposRuleset(),
-                AttrRuleset(),
-                # Verb-Phrase rulesets.
-                RootRuleset(),
-                XcompRuleset(),
-                CcompRuleset(),
-                AdpcompRuleset(),
-                CsubjRuleset(),
-                VmodRuleset(),
-                AdvclRuleset(),
-                RcmodRuleset(),
-                ParataxisRuleset(),
-                # Adjectival-Phrase rulesets
-                AcompRuleset(),
-                AmodRuleset(),
-                # Adverbial-Phrase rulesets
-                AdvmodRuleset(),
-                # Uncategorized rulesets.
-                ConjRuleset(),
-                CompmodJoinRuleset(),
-                DetRuleset(),
-                AdpmodRuleset(),
-                NumRuleset(),
-                WhatRuleset(),
-                CompmodRuleset()]
+ 
+from idd3.rules import universal
