+++ idd3/rules/universal/atomic_emitting_rulesets.py
