+++ idd3/rules/universal/atomic_rulesets.py
