+++ idd3/rules/universal/misc_rulesets.py
@@ -17,9 +17,9 @@
 
 from __future__ import print_function, unicode_literals, division
 from idd3 import Relation, Ruleset
-from idd3.rules.np_rulesets import NounPhraseRuleset
-from idd3.rules.vp_rulesets import VerbPhraseRuleset
-from idd3.rules.adjp_rulesets import AdjectivalPhraseRuleset
+from idd3.rules.universal.np_rulesets import NounPhraseRuleset
+from idd3.rules.universal.vp_rulesets import VerbPhraseRuleset
+from idd3.rules.universal.adjp_rulesets import AdjectivalPhraseRuleset
 
 import logging
 logger = logging.getLogger(__name__)
