+++ idd3/rules/universal/np_rulesets.py
